refactor(A15,A16): split scheduler.rs and merge.rs into focused sub-modules
A15: Split src/scheduler.rs (7768 LOC) into:
  src/scheduler/{mod,worker,metrics,runner,state,signals}.rs

A16: Split src/refresh/merge.rs (3480 LOC) into:
  src/refresh/merge/{mod,columns,conflict,delete,insert,update}.rs

All existing 2060 unit tests pass unchanged. Re-export stubs in mod.rs
preserve the public API surface. Test imports updated to use correct
sub-module paths for the moved helper functions.
