+++ src/scheduler/mod.rs
@@ -54,6 +54,15 @@ use crate::shmem;
 use crate::version;
 use crate::wal_decoder;
 
+pub mod citus;
+pub mod cost;
+pub mod pool;
+pub mod tier;
+
+use citus::drive_distributed_cdc;
+pub use cost::compute_per_db_quota;
+pub use tier::RefreshTier;
+
 // ── SCAL-1 (v0.25.0): Per-backend catalog snapshot cache ─────────────────
 //
 // Caches `StreamTableMeta` rows keyed by the DAG version number from shmem.
@@ -66,16 +75,6 @@ use crate::wal_decoder;
 thread_local! {
     static CATALOG_SNAPSHOT_CACHE: RefCell<Option<(u64, Vec<StreamTableMeta>)>> =
         const { RefCell::new(None) };
-
-    // COORD-14 (v0.34.0): Per-worker consecutive failure counts.
-    //
-    // Key: (pgt_id, "worker_name:port")
-    // Value: number of consecutive poll failures for that worker.
-    //
-    // Reset to 0 on each successful poll; incremented on each failure.
-    // When the count reaches `citus_worker_retry_ticks`, a warning is logged.
-    static CITUS_WORKER_FAILURES: RefCell<HashMap<(i64, String), i32>> =
-        RefCell::new(HashMap::new());
 }
 
 /// SCAL-1: Return active stream tables, using the per-backend snapshot cache.
@@ -140,31 +139,6 @@ pub(crate) fn rebuild_dag_copy_on_write(
     crate::dag::StDag::build_from_catalog(schedule_secs)
 }
 
-/// SCAL-5: Retrieve the configured worker pool size.
-///
-/// Returns 0 (the default) when the persistent worker pool is disabled.
-pub(crate) fn configured_worker_pool_size() -> usize {
-    config::pg_trickle_worker_pool_size().max(0) as usize
-}
-
-/// SCAL-5: Persistent worker pool coordination.
-///
-/// When `pg_trickle.worker_pool_size > 0`, the scheduler maintains a set of
-/// persistent background workers that loop on a work queue rather than being
-/// spawned and de-registered each tick.
-///
-/// The entry is a lightweight marker showing how many pool workers are
-/// currently active. Pool workers are registered via `BackgroundWorkerBuilder`
-/// at extension init time (in `lib.rs`) and communicate via `TICK_WATERMARK_STATE`.
-pub(crate) fn pool_worker_count() -> u32 {
-    shmem::active_worker_count()
-}
-
-/// Check if the persistent worker pool is enabled.
-pub(crate) fn is_pool_enabled() -> bool {
-    configured_worker_pool_size() > 0
-}
-
 /// CACHE-1: Per-backend L0 template cache key check.
 ///
 /// Returns `true` if the local thread-local cache has an entry for `pgt_id`
@@ -177,74 +151,6 @@ pub(crate) fn is_pool_enabled() -> bool {
 pub(crate) fn has_l0_cache_entry(pgt_id: i64) -> bool {
     let current_gen = shmem::current_cache_generation();
     crate::refresh::has_template_cache_entry(pgt_id, current_gen)
-}
-
-// ── G-7: Refresh tier classification ───────────────────────────────────
-
-/// G-7: Refresh tier for tiered scheduling.
-///
-/// Controls the effective schedule multiplier when `pg_trickle.tiered_scheduling`
-/// is enabled. User-assignable via `ALTER STREAM TABLE ... SET (tier = 'warm')`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum RefreshTier {
-    /// Refresh at configured schedule (1× multiplier). Default.
-    #[default]
-    Hot,
-    /// Refresh at 2× configured schedule.
-    Warm,
-    /// Refresh at 10× configured schedule.
-    Cold,
-    /// Skip refresh entirely (manually promoted back to Hot/Warm/Cold).
-    Frozen,
-}
-
-impl RefreshTier {
-    /// Parse from SQL string. Falls back to `Hot` for unknown values.
-    pub fn from_sql_str(s: &str) -> Self {
-        match s.to_lowercase().as_str() {
-            "hot" => RefreshTier::Hot,
-            "warm" => RefreshTier::Warm,
-            "cold" => RefreshTier::Cold,
-            "frozen" => RefreshTier::Frozen,
-            _ => RefreshTier::Hot,
-        }
-    }
-
-    /// SQL representation.
-    pub fn as_str(self) -> &'static str {
-        match self {
-            RefreshTier::Hot => "hot",
-            RefreshTier::Warm => "warm",
-            RefreshTier::Cold => "cold",
-            RefreshTier::Frozen => "frozen",
-        }
-    }
-
-    /// Schedule multiplier for this tier.
-    ///
-    /// Returns `None` for Frozen (skip entirely).
-    pub fn schedule_multiplier(self) -> Option<f64> {
-        match self {
-            RefreshTier::Hot => Some(1.0),
-            RefreshTier::Warm => Some(2.0),
-            RefreshTier::Cold => Some(10.0),
-            RefreshTier::Frozen => None,
-        }
-    }
-
-    /// Validate a tier string. Returns `true` if recognized.
-    pub fn is_valid_str(s: &str) -> bool {
-        matches!(
-            s.to_lowercase().as_str(),
-            "hot" | "warm" | "cold" | "frozen"
-        )
-    }
-}
-
-impl std::fmt::Display for RefreshTier {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.as_str())
-    }
 }
 
 // ── Sub-transaction RAII guard ─────────────────────────────────────────
@@ -326,190 +232,6 @@ pub fn register_launcher_worker() {
         .set_start_time(BgWorkerStartTime::RecoveryFinished)
         .set_restart_time(Some(std::time::Duration::from_secs(5)))
         .load();
-}
-
-/// SCAL-5 (v0.25.0): Spawn persistent pool workers for a database.
-///
-/// Called by the per-database scheduler after connecting to a target DB.
-/// When `pg_trickle.worker_pool_size > 0`, registers N auto-restart BGWs
-/// that loop on the job queue instead of being spawned and de-registered
-/// each tick.  Workers use a 100 ms poll loop and handle `SIGTERM` cleanly.
-///
-/// If `worker_pool_size = 0` (default), this function is a no-op.
-pub(crate) fn spawn_persistent_pool_workers(db_name: &str) {
-    let pool_size = config::pg_trickle_worker_pool_size();
-    if pool_size <= 0 {
-        return;
-    }
-
-    for i in 0..pool_size as u32 {
-        let extra = format!("{db_name}\0{i}");
-        match BackgroundWorkerBuilder::new(&format!("pg_trickle pool worker {i}"))
-            .set_function("pg_trickle_pool_worker_main")
-            .set_library("pg_trickle")
-            .enable_spi_access()
-            .set_extra(&extra)
-            // Pool workers auto-restart on exit (unlike per-job workers).
-            .set_restart_time(Some(std::time::Duration::from_secs(1)))
-            .load_dynamic()
-        {
-            Ok(_) => {
-                log!("pg_trickle: spawned pool worker {i} for db '{db_name}'");
-            }
-            Err(_) => {
-                warning!("pg_trickle: failed to register pool worker {i} for db '{db_name}'");
-            }
-        }
-    }
-}
-
-/// SCAL-5: Entry point for a persistent pool worker.
-///
-/// Loops until `SIGTERM`, polling the job queue and executing QUEUED jobs.
-///
-/// # Safety
-/// Called directly by PostgreSQL as a background worker entry point.
-#[pg_guard]
-#[unsafe(no_mangle)]
-pub extern "C-unwind" fn pg_trickle_pool_worker_main(_arg: pg_sys::Datum) {
-    BackgroundWorker::attach_signal_handlers(SignalWakeFlags::SIGHUP | SignalWakeFlags::SIGTERM);
-
-    let extra = BackgroundWorker::get_extra();
-    // extra = "db_name\0worker_index"
-    let db_name = extra.split('\0').next().unwrap_or("postgres").to_string();
-    let worker_idx: u32 = extra
-        .split('\0')
-        .nth(1)
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(0);
-
-    BackgroundWorker::connect_worker_to_spi(Some(&db_name), None);
-
-    log!(
-        "pg_trickle pool worker {}: started (db='{}')",
-        worker_idx,
-        db_name,
-    );
-
-    // Acquire a worker token (shared with dynamic workers for the cluster budget).
-    let max_workers = config::pg_trickle_max_dynamic_refresh_workers().max(1) as u32;
-    if !shmem::try_acquire_worker_token(max_workers) {
-        log!(
-            "pg_trickle pool worker {}: could not acquire token — cluster budget exhausted, exiting",
-            worker_idx,
-        );
-        return;
-    }
-
-    // Main poll loop: pick up and execute QUEUED jobs.
-    loop {
-        // Check for SIGTERM signal (returns false when signal received).
-        if !BackgroundWorker::wait_latch(Some(std::time::Duration::from_millis(0))) {
-            break;
-        }
-
-        // Try to claim one QUEUED job.
-        let claimed = execute_pool_worker_tick(&db_name, worker_idx);
-
-        if !claimed {
-            // No work available — sleep 100 ms.
-            let ok = BackgroundWorker::wait_latch(Some(std::time::Duration::from_millis(100)));
-            if !ok {
-                // SIGTERM received during sleep.
-                break;
-            }
-        }
-    }
-
-    shmem::release_worker_token();
-    log!(
-        "pg_trickle pool worker {}: exiting (db='{}')",
-        worker_idx,
-        db_name,
-    );
-}
-
-/// SCAL-5: Execute one pool worker tick: claim and run one QUEUED job.
-///
-/// Returns `true` if a job was claimed and executed, `false` if the queue
-/// is empty.
-fn execute_pool_worker_tick(db_name: &str, worker_idx: u32) -> bool {
-    // Claim one QUEUED job for this db.
-    let job_id = match BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        Spi::get_one::<i64>(&format!(
-            "UPDATE pgtrickle.pgt_scheduler_jobs \
-             SET status = 'RUNNING', \
-                 started_at = now(), \
-                 worker_pid = pg_backend_pid() \
-             WHERE job_id = ( \
-               SELECT job_id FROM pgtrickle.pgt_scheduler_jobs \
-               WHERE status = 'QUEUED' \
-                 AND db_name = '{db}' \
-               ORDER BY enqueued_at \
-               LIMIT 1 \
-               FOR UPDATE SKIP LOCKED \
-             ) \
-             RETURNING job_id",
-            db = db_name.replace('\'', "''"),
-        ))
-    })) {
-        Ok(Some(id)) => id,
-        _ => return false,
-    };
-
-    log!(
-        "pg_trickle pool worker {}: executing job_id={}",
-        worker_idx,
-        job_id,
-    );
-
-    // Load the job.
-    let job = match BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        crate::catalog::SchedulerJob::get_by_id(job_id)
-    })) {
-        Ok(Some(j)) => j,
-        _ => return true, // job was claimed, but couldn't load it — skip
-    };
-
-    // Execute the job using the same dispatch as the dynamic refresh worker.
-    let outcome = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        BackgroundWorker::transaction(AssertUnwindSafe(|| -> RefreshOutcome {
-            match job.unit_kind.as_str() {
-                "singleton" => execute_worker_singleton(&job),
-                "atomic_group" => execute_worker_atomic_group(&job, false),
-                "repeatable_read_group" => execute_worker_atomic_group(&job, true),
-                "immediate_closure" => execute_worker_immediate_closure(&job),
-                "cyclic_scc" => execute_worker_cyclic_scc(&job),
-                "fused_chain" => execute_worker_fused_chain(&job),
-                _ => {
-                    warning!(
-                        "pg_trickle pool worker {}: unknown unit_kind '{}' for job {}",
-                        worker_idx,
-                        job.unit_kind,
-                        job_id,
-                    );
-                    RefreshOutcome::PermanentFailure
-                }
-            }
-        }))
-    }));
-
-    let status = match outcome {
-        Ok(RefreshOutcome::Success) => "COMPLETED",
-        _ => "FAILED",
-    };
-
-    // Mark job complete.
-    let _ = BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        Spi::run(&format!(
-            "UPDATE pgtrickle.pgt_scheduler_jobs \
-             SET status = '{status}', \
-                 completed_at = now() \
-             WHERE job_id = {job_id}",
-        ))
-    }));
-
-    true
 }
 
 /// Main entry point for the launcher background worker.
@@ -1960,39 +1682,6 @@ fn is_unit_due(unit: &ExecutionUnit, dag: &StDag) -> bool {
 }
 
 // ── C3-1: Per-database worker quota helpers ───────────────────────────────
-
-/// C3-1: Compute the effective per-database worker quota for this dispatch tick.
-///
-/// When `per_db_quota == 0` (disabled), falls back to `max_concurrent_refreshes`
-/// (the legacy per-coordinator cap, not cluster-aware).
-///
-/// When `per_db_quota > 0`, the base entitlement is `per_db_quota`. If the
-/// cluster has spare capacity (active workers < 80% of `max_cluster`), the
-/// effective quota is increased to `per_db_quota * 3 / 2` to absorb a burst
-/// without wasting idle cluster resources. The burst is reclaimed automatically
-/// within 1 scheduler cycle once global load rises.
-///
-/// Pure logic — extracted for unit-testability.
-pub fn compute_per_db_quota(
-    per_db_quota: i32,
-    max_concurrent_refreshes: i32,
-    max_cluster: u32,
-    current_active: u32,
-) -> u32 {
-    if per_db_quota <= 0 {
-        // C3-1 disabled — fall back to per-coordinator cap (legacy).
-        return max_concurrent_refreshes.max(1) as u32;
-    }
-    let base = per_db_quota.max(1) as u32;
-    // Burst threshold: 80% of cluster capacity.
-    let burst_threshold = ((max_cluster as f64) * 0.8).ceil() as u32;
-    if current_active < burst_threshold {
-        // Spare capacity — allow up to 150% of base quota.
-        (base * 3 / 2).max(base + 1)
-    } else {
-        base
-    }
-}
 
 /// C3-1: Sort an execution unit ready queue by dispatch priority.
 ///
@@ -5583,264 +5272,7 @@ fn last_refresh_row_counts(pgt_id: i64) -> (i64, i64) {
     })
 }
 
-// ── COORD-10/11/12/13/14 (v0.34.0): Distributed CDC pre-refresh ─────────────
-
-/// Drive the distributed CDC pre-refresh step for a stream table that has one
-/// or more sources with `source_placement = 'distributed'`.
-///
-/// Called from [`refresh_single_st`] *before* the local refresh step.  When
-/// Citus is not loaded or the stream table has no distributed sources, this
-/// function returns immediately.
-///
-/// # Behaviour
-///
-/// 1. **COORD-13**: Check topology changes by comparing `pg_dist_node` against
-///    `pgt_worker_slots`.  When the topology has changed, reconcile
-///    `pgt_worker_slots` (drop stale, insert new), mark the stream table
-///    `needs_reinit = true` (triggers FULL refresh), and return early.
-///
-/// 2. **COORD-12**: Acquire a `pgt_st_locks` lease (key = `pgt_<pgt_id>`).
-///    If acquisition fails (another coordinator holds it), return early — the
-///    other coordinator's refresh will handle polling.
-///
-/// 3. **COORD-11**: Call `ensure_worker_slot()` for each entry in
-///    `pgt_worker_slots`.
-///
-/// 4. **COORD-10**: Call `poll_worker_slot_changes()` for each worker slot.
-///    Changes are written into the local change buffer and will be picked up
-///    by the existing differential / full refresh path below.
-///
-/// 5. **COORD-14**: Track consecutive failures per worker.  After
-///    `citus_worker_retry_ticks` failures, emit a `WARNING` to the PostgreSQL
-///    log for operator attention.
-///
-/// Returns the lock key that was acquired (if any), so the caller can release
-/// it after the local refresh completes.
-fn drive_distributed_cdc(pgt_id: i64) -> Option<String> {
-    if !crate::citus::is_citus_loaded() {
-        return None;
-    }
-
-    // Check whether this ST has any distributed sources.
-    let deps = match crate::catalog::StDependency::get_for_st(pgt_id) {
-        Ok(d) => d,
-        Err(e) => {
-            pgrx::debug1!(
-                "[pg_trickle] drive_distributed_cdc: failed to load deps for pgt_id={}: {}",
-                pgt_id,
-                e
-            );
-            return None;
-        }
-    };
-
-    let has_distributed = deps.iter().any(|d| d.source_placement == "distributed");
-    if !has_distributed {
-        return None;
-    }
-
-    // COORD-13: Detect topology changes.
-    if crate::citus::detect_topology_change(pgt_id) {
-        pgrx::log!(
-            "[pg_trickle] COORD-13: topology change detected for pgt_id={} — reconciling worker slots",
-            pgt_id
-        );
-        match crate::citus::reconcile_worker_slots(pgt_id) {
-            Ok(n) => {
-                pgrx::log!(
-                    "[pg_trickle] COORD-13: reconciled {} worker slot entries for pgt_id={}",
-                    n,
-                    pgt_id
-                );
-            }
-            Err(e) => {
-                pgrx::warning!(
-                    "[pg_trickle] COORD-13: reconcile_worker_slots failed for pgt_id={}: {}",
-                    pgt_id,
-                    e
-                );
-            }
-        }
-        // Mark the ST for a full refresh to catch up after topology change.
-        let _ = crate::catalog::StreamTableMeta::mark_for_reinitialize(pgt_id);
-        // Return without polling: the FULL refresh on this tick will catch up.
-        return None;
-    }
-
-    // COORD-12: Acquire pgt_st_locks lease.
-    let lease_ms = config::pg_trickle_citus_st_lock_lease_ms();
-    let lock_key = format!("pgt_{pgt_id}");
-    // Use the local backend PID as the holder so we can recognize our own locks.
-    let holder = format!("scheduler_{}", unsafe { pg_sys::MyProcPid });
-
-    if lease_ms > 0 {
-        match crate::citus::try_acquire_st_lock(&lock_key, &holder, lease_ms) {
-            Ok(true) => {
-                pgrx::debug1!(
-                    "[pg_trickle] COORD-12: acquired pgt_st_locks '{}' ({}ms)",
-                    lock_key,
-                    lease_ms
-                );
-            }
-            Ok(false) => {
-                pgrx::debug1!(
-                    "[pg_trickle] COORD-12: pgt_st_locks '{}' held by another coordinator — skipping",
-                    lock_key
-                );
-                return None;
-            }
-            Err(e) => {
-                pgrx::warning!(
-                    "[pg_trickle] COORD-12: try_acquire_st_lock failed for '{}': {}",
-                    lock_key,
-                    e
-                );
-                return None;
-            }
-        }
-    }
-
-    // Load worker slots.
-    let worker_slots = crate::citus::get_worker_slots_for_st(pgt_id);
-    if worker_slots.is_empty() {
-        // No slots registered yet; reconcile on next tick.
-        if lease_ms > 0 {
-            let _ = crate::citus::release_st_lock(&lock_key, &holder);
-        }
-        return None;
-    }
-
-    let retry_threshold = config::pg_trickle_citus_worker_retry_ticks();
-
-    for slot in &worker_slots {
-        let worker = slot.node_addr();
-        let worker_key = format!("{}:{}", slot.worker_name, slot.worker_port);
-
-        // COORD-11: Ensure the slot exists on the worker.
-        if let Err(e) = crate::citus::ensure_worker_slot(&worker, &slot.slot_name) {
-            pgrx::warning!(
-                "[pg_trickle] COORD-11: ensure_worker_slot failed for {} slot '{}': {}",
-                worker_key,
-                slot.slot_name,
-                e
-            );
-            // Track failure and continue to next worker.
-            record_worker_failure(pgt_id, &worker_key, retry_threshold);
-            continue;
-        }
-
-        // COORD-10: Build source descriptor and poll.
-        // Resolve pk_columns and all columns needed for the change buffer.
-        let pk_columns = match crate::cdc::resolve_pk_columns(slot.source_relid) {
-            Ok(pk) => pk,
-            Err(e) => {
-                pgrx::warning!(
-                    "[pg_trickle] COORD-10: resolve_pk_columns for oid={} failed: {}",
-                    slot.source_relid.to_u32(),
-                    e
-                );
-                record_worker_failure(pgt_id, &worker_key, retry_threshold);
-                continue;
-            }
-        };
-
-        let columns = match crate::cdc::resolve_source_column_defs(slot.source_relid) {
-            Ok(c) => c,
-            Err(e) => {
-                pgrx::warning!(
-                    "[pg_trickle] COORD-10: resolve_source_column_defs for oid={} failed: {}",
-                    slot.source_relid.to_u32(),
-                    e
-                );
-                record_worker_failure(pgt_id, &worker_key, retry_threshold);
-                continue;
-            }
-        };
-
-        // Resolve source schema.table name.
-        let source_qualified_table = crate::citus::stable_name_for_oid(slot.source_relid)
-            .ok()
-            .and_then(|_| {
-                // Build qualified name from catalog.
-                pgrx::Spi::get_one_with_args::<String>(
-                    "SELECT n.nspname || '.' || c.relname \
-                     FROM pg_class c \
-                     JOIN pg_namespace n ON n.oid = c.relnamespace \
-                     WHERE c.oid = $1",
-                    &[slot.source_relid.into()],
-                )
-                .ok()
-                .flatten()
-            })
-            .unwrap_or_else(|| slot.source_relid.to_u32().to_string());
-
-        let change_schema = "pgtrickle_changes";
-        let max_changes: i64 = 10_000;
-
-        let src = crate::citus::WorkerPollSource {
-            change_schema,
-            source_qualified_table: &source_qualified_table,
-            source_oid: slot.source_relid,
-            pk_columns: &pk_columns,
-            columns: &columns,
-        };
-
-        match crate::citus::poll_worker_slot_changes(&worker, &slot.slot_name, max_changes, &src) {
-            Ok(n_changes) => {
-                pgrx::debug1!(
-                    "[pg_trickle] COORD-10: polled {} changes from worker {} slot '{}'",
-                    n_changes,
-                    worker_key,
-                    slot.slot_name
-                );
-                // COORD-14: Reset failure counter on success.
-                reset_worker_failure(pgt_id, &worker_key);
-            }
-            Err(e) => {
-                pgrx::warning!(
-                    "[pg_trickle] COORD-10: poll_worker_slot_changes failed for worker {} slot '{}': {}",
-                    worker_key,
-                    slot.slot_name,
-                    e
-                );
-                record_worker_failure(pgt_id, &worker_key, retry_threshold);
-            }
-        }
-    }
-
-    // Return the lock key so the caller can release it after local refresh.
-    if lease_ms > 0 { Some(lock_key) } else { None }
-}
-
-/// COORD-14: Increment the consecutive failure counter for a worker and emit
-/// an operator warning if the threshold is reached.
-fn record_worker_failure(pgt_id: i64, worker_key: &str, threshold: i32) {
-    CITUS_WORKER_FAILURES.with(|m| {
-        let mut binding = m.borrow_mut();
-        let count = binding
-            .entry((pgt_id, worker_key.to_string()))
-            .or_insert(0);
-        *count += 1;
-        let current = *count;
-        if threshold > 0 && current >= threshold {
-            pgrx::warning!(
-                "[pg_trickle] COORD-14: worker {} has failed {} consecutive times for pgt_id={} \
-                 — stream table flagged in citus_status. Refreshes continue against healthy workers.",
-                worker_key,
-                current,
-                pgt_id
-            );
-        }
-    });
-}
-
-/// COORD-14: Reset the consecutive failure counter for a worker after a
-/// successful poll.
-fn reset_worker_failure(pgt_id: i64, worker_key: &str) {
-    CITUS_WORKER_FAILURES.with(|m| {
-        m.borrow_mut().remove(&(pgt_id, worker_key.to_string()));
-    });
-}
+// ── COORD-10/11/12/13/14 — moved to citus.rs sub-module ──────────────────
 
 /// Refresh a single (non-group) stream table with full retry handling.
 ///
@@ -7401,67 +6833,6 @@ mod tests {
         assert!(!result);
     }
 
-    // ── RefreshTier tests ─────────────────────────────────────────────
-
-    #[test]
-    fn test_refresh_tier_from_sql_str() {
-        assert_eq!(RefreshTier::from_sql_str("hot"), RefreshTier::Hot);
-        assert_eq!(RefreshTier::from_sql_str("WARM"), RefreshTier::Warm);
-        assert_eq!(RefreshTier::from_sql_str("Cold"), RefreshTier::Cold);
-        assert_eq!(RefreshTier::from_sql_str("FROZEN"), RefreshTier::Frozen);
-        assert_eq!(RefreshTier::from_sql_str("unknown"), RefreshTier::Hot);
-        assert_eq!(RefreshTier::from_sql_str(""), RefreshTier::Hot);
-    }
-
-    #[test]
-    fn test_refresh_tier_as_str() {
-        assert_eq!(RefreshTier::Hot.as_str(), "hot");
-        assert_eq!(RefreshTier::Warm.as_str(), "warm");
-        assert_eq!(RefreshTier::Cold.as_str(), "cold");
-        assert_eq!(RefreshTier::Frozen.as_str(), "frozen");
-    }
-
-    #[test]
-    fn test_refresh_tier_schedule_multiplier() {
-        assert_eq!(RefreshTier::Hot.schedule_multiplier(), Some(1.0));
-        assert_eq!(RefreshTier::Warm.schedule_multiplier(), Some(2.0));
-        assert_eq!(RefreshTier::Cold.schedule_multiplier(), Some(10.0));
-        assert_eq!(RefreshTier::Frozen.schedule_multiplier(), None);
-    }
-
-    #[test]
-    fn test_refresh_tier_is_valid_str() {
-        assert!(RefreshTier::is_valid_str("hot"));
-        assert!(RefreshTier::is_valid_str("WARM"));
-        assert!(RefreshTier::is_valid_str("Cold"));
-        assert!(RefreshTier::is_valid_str("FROZEN"));
-        assert!(!RefreshTier::is_valid_str(""));
-        assert!(!RefreshTier::is_valid_str("invalid"));
-    }
-
-    #[test]
-    fn test_refresh_tier_display() {
-        assert_eq!(format!("{}", RefreshTier::Hot), "hot");
-        assert_eq!(format!("{}", RefreshTier::Frozen), "frozen");
-    }
-
-    #[test]
-    fn test_refresh_tier_default() {
-        assert_eq!(RefreshTier::default(), RefreshTier::Hot);
-    }
-
-    #[test]
-    fn test_refresh_tier_roundtrip() {
-        for tier in [
-            RefreshTier::Hot,
-            RefreshTier::Warm,
-            RefreshTier::Cold,
-            RefreshTier::Frozen,
-        ] {
-            assert_eq!(RefreshTier::from_sql_str(tier.as_str()), tier);
-        }
-    }
-
     // ── FUSE unit tests ────────────────────────────────────────────────
 
     #[test]
@@ -7492,60 +6863,6 @@ mod tests {
         assert!(fuse_is_active("auto", Some(100), 0));
         assert!(fuse_is_active("auto", None, 5000));
         assert!(!fuse_is_active("auto", None, 0));
-    }
-
-    // ── C3-1: compute_per_db_quota tests ──────────────────────────────────
-
-    #[test]
-    fn test_quota_disabled_falls_back_to_max_concurrent_refreshes() {
-        // per_db_quota=0 → legacy behaviour, use max_concurrent_refreshes
-        assert_eq!(compute_per_db_quota(0, 4, 8, 0), 4);
-        assert_eq!(compute_per_db_quota(0, 1, 8, 6), 1);
-        // max_concurrent_refreshes=0 is clamped to 1
-        assert_eq!(compute_per_db_quota(0, 0, 8, 0), 1);
-    }
-
-    #[test]
-    fn test_quota_base_case_no_burst_when_cluster_busy() {
-        // active (7) >= 80% of max_cluster (10) = 8 → no burst, return base
-        assert_eq!(compute_per_db_quota(3, 4, 10, 8), 3);
-        assert_eq!(compute_per_db_quota(3, 4, 10, 9), 3);
-        assert_eq!(compute_per_db_quota(3, 4, 10, 10), 3);
-    }
-
-    #[test]
-    fn test_quota_burst_when_cluster_has_spare_capacity() {
-        // active (4) < 80% of max_cluster (10) = 8 → burst to 150% of base
-        // base=4, burst = 4*3/2 = 6
-        assert_eq!(compute_per_db_quota(4, 4, 10, 4), 6);
-        // base=2, burst = 2*3/2 = 3
-        assert_eq!(compute_per_db_quota(2, 4, 10, 0), 3);
-    }
-
-    #[test]
-    fn test_quota_burst_minimum_is_base_plus_one() {
-        // base=1, 1*3/2 = 1 (integer div), but min is base+1=2
-        assert_eq!(compute_per_db_quota(1, 4, 10, 0), 2);
-    }
-
-    #[test]
-    fn test_quota_negative_per_db_quota_falls_back_to_legacy() {
-        assert_eq!(compute_per_db_quota(-1, 3, 10, 0), 3);
-    }
-
-    #[test]
-    fn test_quota_cluster_idle_gives_burst() {
-        // max_cluster=8, 80% threshold = ceil(6.4)=7, active=0 < 7 → burst
-        // base=2, burst=3
-        assert_eq!(compute_per_db_quota(2, 4, 8, 0), 3);
-    }
-
-    #[test]
-    fn test_quota_burst_threshold_at_exactly_80_percent() {
-        // max_cluster=10, ceil(10*0.8)=8, active=7 < 8 → burst allowed
-        assert_eq!(compute_per_db_quota(4, 4, 10, 7), 6);
-        // active=8 == threshold → no burst
-        assert_eq!(compute_per_db_quota(4, 4, 10, 8), 4);
     }
 
     // ── DAG-2: compute_adaptive_poll_ms tests ─────────────────────────────
